pyproject.toml
[project]
name = "ord-schema"
version = "0.3.99"
description = "Schema for the Open Reaction Database"
# url = 'https://github.com/Open-Reaction-Database/ord-schema'"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
# package_data={"ord_schema.proto": ["*.pyi"]}
readme = "README.md"
authors = [
    { name = "hobs", email = "git@totalgood.com" }
]
requires-python = ">=3.12,<3.13"
dependencies = [
    "Werkzeug>=3.0.3",
    "docopt>=0.6.2",
    "inflection>=0.5.1",
    "joblib>=1.0.0",
    "jupyter>=1.1.1",
    "jupyterlab-limit-output>=2.0.0",
    "numpy>=1.26.4",
    "openpyxl>=3.0.5",
    "pandas>=1.0.4",
    "protobuf>=4.22.3,<5",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary,pool]>=3",
    "pygithub>=1.51",
    "python-dateutil>=1.10.0",
    "python-dotenv>=1.2.2",
    "rdkit>=2021.9.5",
    "sqlalchemy>=2",
    "tensorflow>=2.19.1",
    "tqdm>=4.61.2",
]

[project.optional-dependencies]
docs = [
    "Pygments>=2.13.0",
    "Sphinx>=5.3.0",
    "ipython>=8.4.0",
    "sphinx-rtd-theme>=1.1.1",
    "sphinx-tabs>=3.4.0",
]
examples = [
    "glob2>=0.7",
    "matplotlib>=3.3.4",
    "scikit-learn>=0.24.1",
    "tensorflow>=2.4.1",
    "wget>=3.2",
]
tests = [
    "black[jupyter]>=22.3.0",
    "coverage>=5.2.1",
    "pylint>=2.17.4",
    "pytest>=7.1.1",
    "pytest-cov>=3.0.0",
    "pytest-xdist>=3.0.2",
    "pytype>=2023.6.16",
    "testing-postgresql>=1.3.0",
    "treon>=0.1.3",
]


[tool.black]
line-length = 120
extend-exclude = "_pb2.py"

[tool.isort]
skip_glob = ["**/*_pb2.py*"]
line_length = 120
profile = "black"

[tool.pytype]
inputs = ["ord_schema"]
exclude = ["**/*_test.py", "**/test_*.py", "**/*_pb2.py", "build/", "dist/"]
keep_going = true
disable = ["module-attr", "pyi-error"]
jobs = 0

[build-system]
requires = ["uv_build>=0.9.24,<0.10.0"]
build-backend = "uv_build"
